Port to the MEE, and add an MEE HiFiv1 config
This is a bit awkward as it needs to support both MEE-based and direct
implementations.  In an ideal world we'd split out libfemto into its own
library, which would get hooked in via a "-mlibc=femto".  This would
then allow us to avoid the use of a bare toolchain (ie,"-nostartfiles"
and "-nostdlib"), which would make the MEE-based arguments a lot simpler
-- we'd just need to set CFLAGS and LDFLAGS to "-mlibc=femto -menv=mee
-mmachine=sifive-hifive1".

As it currently stands I'm doing something a bit funky, where we use the
MEE's command-line arguments for the link but the existing ones for the
compile.  This is fine for now, but it might beak in the future: for
example, if the MEE decides to use a different ABI than this build
system does.  I also have to add in "-lriscv__mmachine__sifive-hifive1",
which is necessary because we pass in "-nostdlib".

The last wrinkle is that I'm doing the C runtime initialization in C,
which isn't really kosher.  In this case it's probably fine because GCC
is unlikely to rely on an initialized runtime for the functions that are
present, but to be on the safe side this should really be done in
assembly.

Signed-off-by: Palmer Dabbelt <palmer@dabbelt.com>